+++ Confroid/app/src/main/java/fr/uge/confroid/settings/LanguageAdapter.kt
@@ -1,19 +1,23 @@
 package fr.uge.confroid.settings
 
-import android.app.Activity
 import android.content.Context
 import android.content.res.TypedArray
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.BaseAdapter
-import android.widget.ImageView
-import android.widget.TextView
 import fr.uge.confroid.R
-import kotlinx.android.synthetic.main.country_spinner_row.*
 import kotlinx.android.synthetic.main.country_spinner_row.view.*
 
-
+/***
+ * The [BaseAdapter] for the spinner of the languages
+ *
+ * @author Ailton Lopes Mendes
+ * @author Jonathan CHU
+ * @author Fabien LAMBERT--DELAVAQUERIE
+ * @author Akram MALEK
+ * @author Gérald LIN
+ */
 class LanguageAdapter(val context: Context, var languageNameList: Array<String>, var flagList: TypedArray) : BaseAdapter() {
 
     override fun getCount(): Int {
@@ -33,7 +37,7 @@ class LanguageAdapter(val context: Context, var languageNameList: Array<String>,
         val name = languageNameList[position]
 
         return LayoutInflater.from(context).inflate(R.layout.country_spinner_row, null).apply {
-            countryName.text = name.toString()
+            countryName.text = name
             countryFlag.setImageDrawable(flagList.getDrawable(position))
         }
     }
