+++ Confroid/app/src/main/java/fr/uge/confroid/settings/help/fragments/HelpAllVersionsFragment.kt
@@ -7,9 +7,16 @@ import androidx.fragment.app.Fragment
 import fr.uge.confroid.R
 import fr.uge.confroid.utils.ConfroidAnimationUtils
 import kotlinx.android.synthetic.main.fragment_help_all_versions.*
-import kotlinx.android.synthetic.main.fragment_help_app.*
-import kotlinx.android.synthetic.main.fragment_help_setting.*
 
+/***
+ * Will help the user understand better the AllVersionsFragment.
+ *
+ * @author Ailton Lopes Mendes
+ * @author Jonathan CHU
+ * @author Fabien LAMBERT--DELAVAQUERIE
+ * @author Akram MALEK
+ * @author Gérald LIN
+ */
 class HelpAllVersionsFragment : Fragment(R.layout.fragment_help_all_versions) {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
