+++ Confroid/app/src/main/java/fr/uge/confroid/settings/help/fragments/HelpAppFragment.kt
@@ -8,6 +8,15 @@ import fr.uge.confroid.R
 import fr.uge.confroid.utils.ConfroidAnimationUtils
 import kotlinx.android.synthetic.main.fragment_help_app.*
 
+/***
+ * Will help the user understand better the AppsFragment.
+ *
+ * @author Ailton Lopes Mendes
+ * @author Jonathan CHU
+ * @author Fabien LAMBERT--DELAVAQUERIE
+ * @author Akram MALEK
+ * @author Gérald LIN
+ */
 class HelpAppFragment : Fragment(R.layout.fragment_help_app) {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
