+++ Confroid/app/src/main/java/fr/uge/confroid/settings/help/fragments/HelpLoginFragment.kt
@@ -9,7 +9,15 @@ import fr.uge.confroid.R
 import kotlinx.android.synthetic.main.fragment_help_app.*
 import kotlinx.android.synthetic.main.fragment_help_login.*
 
-
+/***
+ * Will help the user understand better the LoginFragment.
+ *
+ * @author Ailton Lopes Mendes
+ * @author Jonathan CHU
+ * @author Fabien LAMBERT--DELAVAQUERIE
+ * @author Akram MALEK
+ * @author Gérald LIN
+ */
 class HelpLoginFragment : Fragment(R.layout.fragment_help_login) {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
