+++ Confroid/app/src/main/java/fr/uge/confroid/settings/help/fragments/HelpMenuFragment.kt
@@ -9,7 +9,15 @@ import fr.uge.confroid.utils.ConfroidAnimationUtils
 import fr.uge.confroid.R
 import kotlinx.android.synthetic.main.fragment_help_menu.*
 
-
+/***
+ * Will help the user understand better the Drawer Menu.
+ *
+ * @author Ailton Lopes Mendes
+ * @author Jonathan CHU
+ * @author Fabien LAMBERT--DELAVAQUERIE
+ * @author Akram MALEK
+ * @author Gérald LIN
+ */
 class HelpMenuFragment : Fragment(R.layout.fragment_help_menu) {
     @RequiresApi(Build.VERSION_CODES.LOLLIPOP)
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
