+++ Confroid/app/src/main/java/fr/uge/confroid/settings/help/fragments/HelpRegisterFragment.kt
@@ -8,6 +8,15 @@ import fr.uge.confroid.utils.ConfroidAnimationUtils
 import fr.uge.confroid.R
 import kotlinx.android.synthetic.main.fragment_help_register.*
 
+/***
+ * Will help the user understand better the RegisterFragment.
+ *
+ * @author Ailton Lopes Mendes
+ * @author Jonathan CHU
+ * @author Fabien LAMBERT--DELAVAQUERIE
+ * @author Akram MALEK
+ * @author Gérald LIN
+ */
 class HelpRegisterFragment : Fragment(R.layout.fragment_help_register) {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
