+++ Confroid/app/src/main/java/fr/uge/confroid/storageprovider/MyProvider.kt
@@ -1,20 +1,28 @@
 package fr.uge.confroid.storageprovider
 
 import android.content.Context
-import android.content.Intent
-import android.os.Build
-import android.os.Environment
 import android.util.Log
-import androidx.annotation.RequiresApi
-import androidx.core.content.FileProvider.getUriForFile
 import java.io.File
 import java.io.FileOutputStream
 import java.io.IOException
 
-
+/***
+ * The provider to write files and get them from the memory of the phone.
+ *
+ * @author Ailton Lopes Mendes
+ * @author Jonathan CHU
+ * @author Fabien LAMBERT--DELAVAQUERIE
+ * @author Akram MALEK
+ * @author Gérald LIN
+ */
 object MyProvider {
 
-
+    /***
+     * Will write the [content] given on the [File].
+     * @param context the [Context].
+     * @param fileName the name of the [File].
+     * @param content the data to overwrite on the [File].
+     */
     fun writeFile(context: Context, fileName : String, content : ByteArray) : File {
         val file = StorageUtils.createFile(fileName, context.filesDir.absolutePath)
         Log.i("file path", "path $file")
@@ -29,6 +37,10 @@ object MyProvider {
         return file
     }
 
+    /***
+     * @param context the [Context].
+     * @return Will return an [Array] of [File].
+     */
     fun getFiles(context: Context) : Array<File> {
         return StorageUtils.getFiles(context.filesDir.absolutePath)
     }
