+++ Confroid/app/src/main/java/fr/uge/confroid/storageprovider/StorageUtils.kt
@@ -1,11 +1,25 @@
 package fr.uge.confroid.storageprovider
 
-import android.content.Context
 import android.util.Log
 import java.io.File
 
+/***
+ * Will deal with the features of an file by creating or getting.
+ *
+ * @author Ailton Lopes Mendes
+ * @author Jonathan CHU
+ * @author Fabien LAMBERT--DELAVAQUERIE
+ * @author Akram MALEK
+ * @author Gérald LIN
+ */
 object StorageUtils {
 
+    /***
+     * Will create a file. If the file already exists, it will delete the [File] first then create the [File].
+     * @param fileName the [File] name.
+     * @param folderName the folder name.
+     * @return a [File]
+     */
     fun createFile(
         fileName: String?,
         folderName: String?
@@ -25,6 +39,11 @@ object StorageUtils {
         return file
     }
 
+    /***
+     * Will return all the [File] in a folder.
+     * @param folderName the folder name.
+     * @return An [Array] of [File].
+     */
     fun getFiles(folderName: String?) : Array<File> {
         val root = File(folderName)
         if (!root.exists() && !root.isDirectory) {
