+++ Confroid/app/src/main/java/fr/uge/confroid/utils/CustomDiffUtil.kt
@@ -1,8 +1,16 @@
 package fr.uge.confroid.utils
 
 import androidx.recyclerview.widget.DiffUtil
-import androidx.recyclerview.widget.RecyclerView
 
+/***
+ * Will generate a custom [DiffUtil].
+ *
+ * @author Ailton Lopes Mendes
+ * @author Jonathan CHU
+ * @author Fabien LAMBERT--DELAVAQUERIE
+ * @author Akram MALEK
+ * @author Gérald LIN
+ */
 class CustomDiffUtil(private val old: List<*>, private val current: List<*>) : DiffUtil.Callback() {
     override fun getOldListSize(): Int {
         return old.size
