+++ Confroid/app/src/main/java/fr/uge/confroid/web/FileAttributes.kt
@@ -1,3 +1,12 @@
 package fr.uge.confroid.web
 
+/***
+ * Data class of the attributes of a File.
+ *
+ * @author Ailton Lopes Mendes
+ * @author Jonathan CHU
+ * @author Fabien LAMBERT--DELAVAQUERIE
+ * @author Akram MALEK
+ * @author Gérald LIN
+ */
 data class FileAttributes(val name : String, val url : String)
