+++ Confroid/app/src/main/java/fr/uge/confroid/web/LoginFragment.kt
@@ -18,6 +18,7 @@ import androidx.work.NetworkType
 import androidx.work.PeriodicWorkRequestBuilder
 import androidx.work.WorkManager
 import fr.uge.confroid.R
+import fr.uge.confroid.web.LoginRequest.logAction
 import fr.uge.confroid.worker.UploadWorker
 import kotlinx.android.synthetic.main.fragment_login.*
 import java.util.concurrent.TimeUnit
@@ -94,36 +95,9 @@ class LoginFragment : Fragment(R.layout.fragment_login) {
             //Intent(activity, MainActivity::class.java).apply { startActivity(this) }
             viewModel.selectItem(this)
             navController.navigate(R.id.action_loginFragment_to_appFragment)
-            logAction()
+            logAction(requireContext())
             //mainDrawerLayout.invalidate()
 
         }
-    }
-
-    private fun logAction() {
-        if (WebSharedPreferences.getInstance(requireContext()).isLoggedIn()) {
-
-            val user = WebSharedPreferences.getInstance(requireContext()).getUser()
-            LoginRequest.request(requireContext(), user.username, user.password) {}
-            work()
-
-            Log.i("shared user", user.toString())
-
-        }
-    }
-
-    private fun work() {
-        val constraints = Constraints.Builder()
-            .setRequiresBatteryNotLow(true)
-            .setRequiredNetworkType(NetworkType.CONNECTED)
-            .setRequiresStorageNotLow(true)
-            .build()
-
-        val work = PeriodicWorkRequestBuilder<UploadWorker>(20, TimeUnit.MINUTES)
-            .setConstraints(constraints)
-            .build()
-
-        val workManager = WorkManager.getInstance(requireContext())
-        workManager.enqueue(work)
     }
 }
