--- ConfroidWebServices/resources/config-uploads/4/Calculator_1.txt
@@ -1,1 +1,0 @@
-zMW/Sa7tVhNH3K/2Ex/rTxAUZbxXJ4BR2KAZSjoBiDBJ7aSB0ZUku7m51pR1Tf67clLZVBKttozNm6fPPwOz94K8Ejq630VAyCoWZJdoyycXThxOF1AC3+zcvWbiG/j8Osr/iyB0A5vGFkKiJfYChamRd54LysTIAZ/Q5zaoh8Q=
