--- ConfroidWebServices/resources/config-uploads/4/Shop_3.txt
@@ -1,1 +1,0 @@
-GXtQgrJaUDw5N+1ynF03chI6p4oUqtCQyF6NgZhhFC9eN/xvEJXY3m83KhZAfjH+/xEjMymXY5+iu8J3egq4Wg==
