--- ConfroidWebServices/resources/config-uploads/4/Shop_5.txt
@@ -1,1 +1,0 @@
-Bs5Rj2MDvcxrgzbPa0vb7d7TiuKVKLahU+5nuN4RvDDEaIfwY00qOA4t08NzGVFLOswRs/bHk0W8SVDT1/ueNjNnlLzO+fE3Wae9OsjKQB2QUiqlf54bESB5meW6UwtMAQXULW/pUPztYwKIn83qsnBZVqqgTh0JTBtrAKMvZuQzZ5S8zvnxN1mnvTrIykAdRcpZmHreuJUwRu3UhQ1n2a3WOXdPQQMURSckmaclohphYhDjQ8Q5KLJdOnJRNpuc8eHU54NfmM4S5CoM0TxevQLjWpzEmMiAg6nnlbENj+7cOU/o3vvcHMFgXkxuEGJ8Ex0WqT7IXAzzye1Q8j5eOIrZXWCZwAu8W7tEK1Pj07hR+sW1MjA3TW9suDQYM/8kgWdHFPfTbClwrZwxPeLHraRnrdNGWFc1FxLNWBwRy2ANQiro5EcbuHy7Xhk0vKwSR/6RUAI4/2lyZkkSVxbrj2L/nLqDmd1bz4XeCp4997f4cREEpiBtIl48ChtJx7yX
